LPS-79876 Move verifyUuid logic added by LPS-29879 to an upgrade process
