+++ portal-impl/src/com/liferay/portal/upgrade/UpgradeProcess_6_2_0.java
@@ -44,6 +44,7 @@ import com.liferay.portal.upgrade.v6_2_0.UpgradeSchema;
 import com.liferay.portal.upgrade.v6_2_0.UpgradeSearch;
 import com.liferay.portal.upgrade.v6_2_0.UpgradeSocial;
 import com.liferay.portal.upgrade.v6_2_0.UpgradeUser;
+import com.liferay.portal.upgrade.v6_2_0.UpgradeUuid;
 import com.liferay.portal.upgrade.v6_2_0.UpgradeWiki;
 import com.liferay.portal.upgrade.v6_2_0.UpgradeWikiAttachments;
 import com.liferay.portal.verify.VerifyUUID;
@@ -108,6 +109,7 @@ public class UpgradeProcess_6_2_0 extends Pre7UpgradeProcess {
 		upgrade(new UpgradeSocial());
 		upgrade(new UpgradeSQLServer());
 		upgrade(new UpgradeUser());
+		upgrade(new UpgradeUuid());
 		upgrade(new UpgradeWiki());
 		upgrade(new UpgradeWikiAttachments());
